requirements.txt

flask==3.0.2
flask-socketio==5.3.6
flask-sqlalchemy==3.1.1
python-socketio==5.11.0
websockets==12.0
websocket-client==1.6.4
assemblyai==0.42.1
google-generativeai==0.8.3
pydantic==2.9.2
pydantic-core==2.23.4
requests==2.32.3
werkzeug==3.0.4
sqlalchemy==2.0.25
gunicorn==21.2.0
gtts==2.5.1
email-validator==2.1.1
psycopg2-binary==2.9.9
socketio
html2text==2020.1.16
PyPDF2==3.0.1
python-docx==1.1.0
python-dotenv==1.0.0
aiohttp==3.9.1
asyncio-mqtt==0.16.1
duckduckgo-search==4.1.1
beautifulsoup4==4.12.2
lxml==4.9.3
openpyxl==3.1.2
python-magic==0.4.27
chardet==5.2.0
google-generativeai
